replaced vector and rect with qt versions

+++ 3rd_party/oxygine-framework/oxygine/actor/slidingsprite.cpp
@@ -29,12 +29,12 @@ namespace oxygine
         }
     }
 
-    float SlidingSprite::getSpeedX() const
+    qint32 SlidingSprite::getSpeedX() const
     {
         return m_speedX;
     }
 
-    void SlidingSprite::setSpeedX(float speed)
+    void SlidingSprite::setSpeedX(qint32 speed)
     {
         if (!m_locked)
         {
@@ -42,12 +42,12 @@ namespace oxygine
         }
     }
 
-    float SlidingSprite::getSpeedY() const
+    qint32 SlidingSprite::getSpeedY() const
     {
         return m_speedY;
     }
 
-    void SlidingSprite::setSpeedY(float speed)
+    void SlidingSprite::setSpeedY(qint32 speed)
     {
         if (!m_locked)
         {
