+++ ingamescriptsupport/conditions/scriptconditionisco.cpp
@@ -157,7 +157,7 @@ void ScriptConditionIsCo::showEditCondition(spScriptEditor pScriptEditor)
         }
         oxygine::spSprite pSprite = oxygine::spSprite::create();
         pSprite->setResAnim(pAnim);
-        pSprite->setScale(pAnim->getWidth() / 32.0f);
+        pSprite->setScale(static_cast<float>(pAnim->getWidth()) / 32.0f);
         pSprite->setSize(pAnim->getSize());
         return pSprite;
     };
