+++ menue/basegamemenu.h
@@ -41,7 +41,7 @@ public:
     virtual Player* getCurrentViewPlayer();
     void registerAtInterpreter(QString name = "game");
 signals:
-    void sigMouseWheel(float direction);
+    void sigMouseWheel(qint32 direction);
     void sigRightClick(qint32 x, qint32 y);
     void sigRightClickDown(qint32 x, qint32 y);
     void sigRightClickUp(qint32 x, qint32 y);
